+++ app/i18next-options.ts
