+++ app/i18next.server.ts
@@ -1,7 +1,7 @@
 import Backend from "i18next-fs-backend";
 import { resolve } from "node:path";
 import { RemixI18Next } from "remix-i18next";
-import i18nextOptions from "./i18nextOptions";
+import i18nextOptions from "./i18next-options";
 
 let i18next: RemixI18Next = new RemixI18Next({
   detection: {
